requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0  # For EfficientNet models

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics & Evaluation
scikit-learn>=1.3.0

# Progress Bars
tqdm>=4.65.0

# Experiment Tracking (optional)
# wandb>=0.15.0
# tensorboard>=2.13.0

# Utilities
pyyaml>=6.0
